refactor(auth): move session logic to entities layer
- Move SessionProvider and session hooks from features/auth to entities/session
- Update import paths across components and pages
- Maintain clean separation between auth features and session entities
- Remove session exports from auth feature index

This change improves the architecture by properly separating session
management (entity) from authentication features (UI components).
