+++ features/auth/index.ts
@@ -1,5 +1,3 @@
-export { RoleEnum } from '@/services/users';
-export * from '@/entities/session';
 export * from './ui/protected-component';
 export * from './ui/auth-guard-button';
 export * from './ui/login-modal';
