+++ pages/_app.jsx
@@ -7,7 +7,8 @@ import { Toaster as SonnerToaster } from '@/shared/ui/sonner';
 import { useRouter } from 'next/router';
 import Script from 'next/script';
 import Head from 'next/head';
-import { SessionProvider, LoginModal } from '@/features/auth';
+import { LoginModal } from '@/features/auth';
+import { SessionProvider } from '@/entities/session';
 import { DialogProvider } from '@/contexts/Dialog';
 import { fetcher } from '@/shared/lib/http';
 import getBaseLayout from '@/layout/core/getBaseLayout';
